requirements.txt
onnx
onnxruntime
opencv-python
numpy
scikit-image
faiss-cpu
rknn-toolkit2
opencv_contrib_python